Removed unused code
